% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram.R
\name{bigram}
\alias{bigram}
\title{Twitter Bi-Grams}
\usage{
bigram(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A tribble.
}
\description{
Determines and displays the text Bi-Grams within the Twitter data in sequence from the most used to the least used.  A Bi-Gram is a combination of two consecutive words.
}
\examples{
\dontrun{
library(saotd)
data <- raw_tweets
TD_Bigram <- bigram(DataFrame = data)
TD_Bigram
}                
}
