% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bing.R
\docType{data}
\name{Bing}
\alias{Bing}
\title{Bing Lexicon Dictionary}
\format{A \code{tribble} with 6,788 rows and 2 variables

\describe{
 \item{Token}{A word}
 \item{Sentiment}{The determined sentiment of a Token.  Either "positive" or "negative".}
 }}
\usage{
data(Bing)
}
\description{
Dataset containing the Bing Sentiment Lexicon Dictionary
}
\keyword{datasets}
