% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BoxPlot}
\alias{BoxPlot}
\title{Twitter Data Box Plot}
\usage{
BoxPlot(DataFrameTidyScores, HT_Topic)
}
\arguments{
\item{DataFrameTidyScores}{DataFrame of Twitter Data that has been tidy'd and scored.}

\item{HT_Topic}{If using hashtag data select:  "hashtag".  If using topic data select:  "topic".}
}
\value{
A ggplot box plot.
}
\description{
Displays the distribution scores of either hashtag or topic Twitter data.
}
\examples{
\dontrun{
library(saotd)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                    HT_Topic = "hashtag") 
ht_box <- Boxplot(DataFrameTidyScores = score_data,
                  HT_Topic = "hashtag")
ht_box
               
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                    HT_Topic = "topic") 
topic_box <- BoxPlot(DataFrameTidyScores = score_data,
                     HT_Topic = "topic") 
topic_box                    
}
}
