% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Corpus.Distribution}
\alias{Corpus.Distribution}
\title{Twitter Corpus Distribution}
\usage{
Corpus.Distribution(DataFrameTidyScores, color = "black", fill = "white")
}
\arguments{
\item{DataFrameTidyScores}{DataFrame of Twitter Data that has been tidy'd and scored.}

\item{color}{The user selected color to highlight the bins.}

\item{fill}{The interior color of the bins.}
}
\value{
A ggplot.
}
\description{
Determines the scores distribution for the entire Twitter data corpus.
}
\examples{
\dontrun{ 
library(SAoTD)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                    HT_Topic = "hashtag") 
Corp_Dist <- Corpus.Distribution(DataFrameTidyScores = score_data,
                                 color = "black", 
                                 fill = "white")
Corp_Dist
}
}
