% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Max.Scores}
\alias{Max.Scores}
\title{Twitter Data Maximum Scores}
\usage{
Max.Scores(DataFrameTidyScores, HT_Topic, HT_Topic_Selection = NULL)
}
\arguments{
\item{DataFrameTidyScores}{DataFrame of Twitter Data that has been tidy'd and scored.}

\item{HT_Topic}{If using hashtag data select:  "hashtag".  If using topic data select:  "topic".}

\item{HT_Topic_Selection}{THe hashtag or topic to be investigated.  NULL will find min across entire dataframe.}
}
\value{
A Tidy DataFrame.
}
\description{
Determines the Maximum scores for either the entire dataset or the Maximum scores associated with a hashtag or topic analysis.
}
\examples{
\dontrun{
library(SAoTD)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                     HT_Topic = "hashtag")
min_scores <- Max.Scores(DataFrameTidyScores = score_data, 
                        HT_Topic = "hashtag")
                            
data <- twitter_data
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                     HT_Topic = "hashtag")
min_scores <- Max.Scores(DataFrameTidyScores = score_data, 
                         HT_Topic = "hashtag",
                         HT_Topic_Selection = "icecream")
}
}
