% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Min.Scores}
\alias{Min.Scores}
\title{Twitter Data Minimum Scores}
\usage{
Min.Scores(DataFrameTidyScores, HT_Topic, HT_Topic_Seletion = NULL)
}
\arguments{
\item{DataFrameTidyScores}{DataFrame of Twitter Data that has been tidy'd and scored.}

\item{HT_Topic}{If using hashtag data select:  "hashtag".  If using topic data select:  "topic".}

\item{HT_Topic_Selection}{THe hashtag or topic to be investigated.  NULL will find min across entire dataframe.}
}
\value{
A Tidy DataFrame.
}
\description{
Determines the minimum scores for either the entire dataset or the minimum scores associated with a hashtag or topic analysis.
}
\examples{
library(SAoTD)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                    HT_Topic = "hashtag")
min_scores <- Min.Scores(DataFrameTidyScores = score_data, 
                         HT_Topic = "hashtag")
                            
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                     HT_Topic = "hashtag")
min_scores <- Min.Scores(DataFrameTidyScores = score_data, 
                         HT_Topic = "hashtag",
                         HT_Topic_Selection = "icecream")

}
