% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PosNeg.Words}
\alias{PosNeg.Words}
\title{Twitter Positive and Negative Words}
\usage{
PosNeg.Words(DataFrameTidy, num_words, filterword = NULL)
}
\arguments{
\item{DataFrameTidy}{DataFrame of Twitter Data that has been tidy'd.}

\item{num_words}{Desired number of words to be returned.}

\item{filterword}{Word or words to be removed}
}
\value{
A ggplot
}
\description{
Determines and displays the most positive and negative words within the twitter data.
}
\examples{
\dontrun{
library(SAoTD)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
posneg <- PosNeg.Words(DataFrameTidy = tidy_data,
                       n = 10)
posneg
                          
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
posneg <- PosNeg.Words(DataFrameTidy = tidy_data,
                       n = 10,
                       filterword = "fail")
posneg
                          
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
posneg <- PosNeg.Words(DataFrameTidy = tidy_data,
                       n = 10,
                       filterword = c("fail", "urgent"))            
posneg
}
}
