% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_my_app}
\alias{run_my_app}
\title{Runs an app in my package}
\usage{
run_my_app(app_name, ...)
}
\arguments{
\item{app_name}{character string for a directory in this package}

\item{...}{Additional options passed to shinyAppDir}
}
\value{
A printed shiny app
}
\description{
Runs a shiny app
}
\examples{
\dontrun{run_my_app('myfirstapp')}
}
