% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{Stop Words}
\format{A \code{tribble} with 1,149 rows and 1 variables

\describe{
 \item{word}{List of stop words.}
 }}
\usage{
data(stop_words)
}
\description{
Stop Words Dictionary
}
\keyword{datasets}
