% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00. Functions.R
\name{TD.Tidy}
\alias{TD.Tidy}
\title{Tidy Twitter Data}
\usage{
TD.Tidy(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data}
}
\value{
A Tidy DataFrame
}
\description{
Function to Tidy Twitter Data and remove all emoticons whilie maintaiing actual tweet
}
\examples{
TD.Tidy(DataFrame = newdata)
}
