% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Tidy}
\alias{Tidy}
\title{Tidy Twitter Data}
\usage{
Tidy(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A Tidy DataFrame.
}
\description{
Function to Tidy Twitter Data and remove all emoticons whilie maintaiing actual tweet.
}
\examples{
library(SAoTD)
data <- raw_tweets
tidy_data <- TD.Tidy(DataFrame = data)
tidy_data

}
