% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_acquire.R
\name{tweet_acquire}
\alias{tweet_acquire}
\title{Acquire Twitter Tweets}
\usage{
tweet_acquire(
  twitter_app,
  consumer_api_key,
  consumer_api_secret_key,
  access_token,
  access_token_secret,
  query,
  num_tweets,
  distinct = TRUE
)
}
\arguments{
\item{twitter_app}{The name of user created Twitter Application.}

\item{consumer_api_key}{Twitter Application management consumer API key.}

\item{consumer_api_secret_key}{Twitter Application management consumer API 
secret key.  Application must have \code{Read and write} access level and
\code{Callback URL} of \code{http://127.0.0.1:1410}.}

\item{access_token}{Twitter Application management access token 
(apps.twitter.com).}

\item{query}{A single query or a list of queries the user has specified.
Character string, not to exceed 500 characters.  To search for tweets 
containing at least one of multiple possible terms, separate each search 
term with spaces and "OR" (in caps).  For example, the search \code{q =
"data science"} looks for tweets containing both "data" and "science" 
located anywhere in the tweets and in any order.  When "OR" is entered 
between search terms, \code{query = "data OR science"}, Twitter's REST API 
should return any tweet that contains either "data" or "science."}

\item{num_tweets}{Number of Tweets to be acquired per each hashtag.}

\item{distinct}{Logical.  If distinct = TRUE, the function removes multiple 
Tweets that originate from the same Twitter id at the exact same time.}

\item{access_secret}{Twitter Application management access secret token 
(apps.twitter.com).}
}
\value{
A DataFrame with tweets and meta data.
}
\description{
Function will enable a user to access the Twitter API through 
  the [Twitter Developers Account](https://dev.twitter.com/) site.  Once a 
  user has a Twitter developers account and has received their individual 
  consumer key, consumer secret key, access token, and access secret they can 
  acquire Tweets based on a list of hashtags and a requested number of 
  entries per query.
}
\examples{
\donttest{
twitter_app <- "super_app"
consumer_api_key <- "XXXXXXXXXXXXXXXXXXXXXXXXX"
consumer_api_secret_key <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
access_token <- "XXXXXXXXXXXXXXXXXX-XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
access_token_secret <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"

Or the Twitter API keys and tokens can be saved in an .Renviron file.  If 
using a .Renviron the 

tweets <- tweet_acquire(twitter_app = twitter_app,
                        consumer_api_key = consumer_api_key,
                        consumer_api_secret_key = consumer_api_secret_key,
                        access_token = access_token,
                        access_token_secret = access_token_secret,
                        query = "#icecream",
                        num_tweets = 1000,
                        distinct = TRUE)
}
}
