% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_tidy.R
\name{tweet_tidy}
\alias{tweet_tidy}
\title{Tidy Twitter Data}
\usage{
tweet_tidy(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A Tidy tibble.
}
\description{
Function to Tidy Twitter Data.  This function will remove a 
  significant amount of the original twitter metadata, as it is not needed 
  to determine the sentiment fo the tweets. This function will remove all 
  emoticons, punctuation, weblinks while maintaining actual Tweet text.
}
\examples{
\donttest{
library(saotd)

data <- raw_tweets
tidy_data <- tweet_tidy(DataFrame = data)
tidy_data

}
}
