% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{unigram}
\alias{unigram}
\title{Twitter Uni-Grams}
\usage{
unigram(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A tribble.
}
\description{
Determines and displays the text Uni-Grams within the Twitter data in sequence from the most used to the least used.  A Uni-Gram is a single word.
}
\examples{
\dontrun{
library(saotd)
data <- raw_tweets
TD_Unigram <- unigram(DataFrame = data)
TD_Unigram
}             
}
