% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ViolinPlot}
\alias{ViolinPlot}
\title{Twitter Data Violin Plot}
\usage{
ViolinPlot(DataFrameTidyScores, HT_Topic)
}
\arguments{
\item{DataFrameTidyScores}{DataFrame of Twitter Data that has been tidy'd and scored.}

\item{HT_Topic}{If using hashtag data select:  "hashtag".  If using topic data select:  "topic".}
}
\value{
A ggplot violin plot.
}
\description{
Displays the distribution scores of either hashtag or topic Twitter data.
}
\examples{
\dontrun{
library(SAoTD)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                     HT_Topic = "hashtag") 
ht_violin <- ViolinPlot(DataFrameTidyScores = score_data,
                        HT_Topic = "hashtag")
ht_violin
               
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                     HT_Topic = "topic") 
topic_violin <- ViolinPlot(DataFrameTidyScores = score_data,
                           HT_Topic = "topic") 
topic_violin                    
}
}
