% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Word.Corr}
\alias{Word.Corr}
\title{Twitter Word Correlations}
\usage{
Word.Corr(DataFrameTidy, number, sort = TRUE)
}
\arguments{
\item{DataFrameTidy}{DataFrame of Twitter Data that has been tidy'd.}

\item{number}{The number of word instances to be included.}

\item{sort}{Rank order the results from most to least correlated.}
}
\value{
A tribble
}
\description{
The word correlation displays the mutual relationship between words.
}
\examples{
\dontrun{ 
library(SAoTD)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
TD_Word_Corr <- Word.Corr(DataFrameTidy = tidy_data, 
                          number = 500,
                          sort = TRUE)

TD_Word_Corr
  }                        
}
