% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_corr.R
\name{word_corr}
\alias{word_corr}
\title{Twitter Word Correlations}
\usage{
word_corr(DataFrameTidy, number, sort = TRUE)
}
\arguments{
\item{DataFrameTidy}{DataFrame of Twitter Data that has been tidy'd.}

\item{number}{The number of word instances to be included.}

\item{sort}{Rank order the results from most to least correlated.}
}
\value{
A Tibble.
}
\description{
The word correlation displays the mutual relationship between words.
}
\examples{
\donttest{
library(saotd)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
TD_Word_Corr <- word_corr(DataFrameTidy = tidy_data, 
                          number = 500,
                          sort = TRUE)

TD_Word_Corr
}                    
}
