% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Bigram}
\alias{Bigram}
\title{Twitter Bi-Grams}
\usage{
Bigram(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A tribble.
}
\description{
Determines and displays the text Bi-Grams within the Twitter data in sequence from the most used to the least used.  A Bi-Gram is a combination of two consecutive words.
}
\examples{
library(SAoTD)
data <- raw_tweets
TD_Bigram <- Bigram(DataFrame = data)
TD_Bigram
                          
}
