% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_tweets}
\alias{raw_tweets}
\title{Twitter Data Set}
\format{A \code{tribble} with 14,483 rows and 6 variables

\describe{
 \item{id}{ID of this status.}
 \item{hashtag}{Hashtag that the individual tweet was acquired from.}
 \item{screenName}{Screen name of the user who posted this status.}
 \item{text}{The text of the status.}
 \item{created}{When this status was created.}
 \item{key}{Unique key based on the tweets originators user id and the created date time group.}
 }}
\usage{
data(raw_tweets)
}
\description{
Dataset from a [Twitter US Airline Sentiment](https://www.kaggle.com/crowdflower/twitter-airline-sentimen) Kaggle competition, from December 2017.  The dataset contains 14,487 tweets from 6 different hashtags (2,604 x #American, 2,220 x #Delta, 2,420 x #Southwest, 3,822 x #United, 2,913 x #US Airways, 504 x #Virgin America).
}
\keyword{datasets}
