% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigram.R
\name{trigram}
\alias{trigram}
\title{Twitter Tri-Grams}
\usage{
trigram(DataFrame)
}
\arguments{
\item{DataFrame}{Data Frame of Twitter Data.}
}
\value{
A tribble.
}
\description{
Determines and displays the text Tri-Grams within the Twitter 
  data in sequence from the most used to the least used.  A Tri-Gram is a 
  combination of three consecutive words.
}
\examples{
\donttest{
library(saotd)
data <- raw_tweets
TD_Trigram <- trigram(DataFrame = data)
TD_Trigram
}
}
