% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Trigram}
\alias{Trigram}
\title{Twitter Tri-Grams}
\usage{
Trigram(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A tribble.
}
\description{
Determines and displays the text Tri-Grams within the Twitter data in sequence from the most used to the least used.  A Tri-Gram is a combination of three consecutive words.
}
\examples{
\dontrun{
library(SAoTD)
data <- raw_tweets
TD_Trigram <- Trigram(DataFrame = data)
TD_Trigram
}
}
