% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Tweet.Topics}
\alias{Tweet.Topics}
\title{Tweet Topics}
\usage{
Tweet.Topics(DataFrame, clusters, method = "Gibbs", set_seed = 1234,
  num_terms = 10)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}

\item{set_seed}{Seed for reproducable results.}

\item{num_cores}{The number of CPU cores to processes models simultaneously (2L for dual core processor).}

\item{min_clusters}{Lower range for the number of clusters.}

\item{max_clusters}{Upper range for the number of clusters.}

\item{skip}{Integer; The number of clusters to skip between entries.}
}
\value{
Returns LDA topics.
}
\description{
Determines the Latent topics within a dataframe by using Latent Dirichlet Allocation (LDA) model parameters.  
Uses the `ldatuning` package and outputs an ldatuning plot.  
Prepares tweet text, creates DTM, conducts LDA, display data terms associated with each topic.
}
\examples{
library(SAoTD)
data <- twitter_data
LDA_data <- Tweet.Topics(DataFrame = data,
                         clusters = 8,
                         method = "Gibbs",
                         set_seed = 1234,
                         num_terms = 10)

LDA_data

}
