% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_scores.R
\name{tweet_scores}
\alias{tweet_scores}
\title{Score Tidy Twitter Data}
\usage{
tweet_scores(DataFrameTidy, HT_Topic)
}
\arguments{
\item{DataFrameTidy}{Data Frame of Twitter Data that has been tidy'd.}

\item{HT_Topic}{If using hashtag data select:  "hashtag".  If using topic 
data select:  "topic"}
}
\value{
A Scored DataFrame.
}
\description{
Function to Calculate Sentiment Scores that will account for 
  sentiment by hashtag or topic.
}
\examples{
\dontrun{
library(saotd)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- tweet_scores(DataFrameTidy = tidy_data,
                           HT_Topic = "hashtag")
score_data
}
}
