% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_tidy.R
\name{tweet_tidy}
\alias{tweet_tidy}
\title{Tidy Twitter Data}
\usage{
tweet_tidy(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A Tidy DataFrame.
}
\description{
Function to Tidy Twitter Data and remove all emoticons, punctuation, weblinks while maintaiing actual Tweet.
}
\examples{
\donttest{
library(saotd)
data <- raw_tweets
tidy_data <- tweet_tidy(DataFrame = data)
tidy_data
}
}
