% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigram.R
\name{unigram}
\alias{unigram}
\title{Twitter Uni-Grams}
\usage{
unigram(DataFrame)
}
\arguments{
\item{DataFrame}{Data Frame of Twitter Data.}
}
\value{
A tibble.
}
\description{
Determines and displays the text Uni-Grams within the Twitter 
  data in sequence from the most used to the least used.  A Uni-Gram is a 
  single word.
}
\examples{
\dontrun{
library(saotd)
data <- raw_tweets
TD_Unigram <- unigram(DataFrame = data)
TD_Unigram
}             
}
