% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Unigram}
\alias{Unigram}
\title{Twitter Uni-Grams}
\usage{
Unigram(DataFrame)
}
\arguments{
\item{DataFrame}{DataFrame of Twitter Data.}
}
\value{
A tribble.
}
\description{
Determines and displays the text Uni-Grams within the Twitter data in sequence from the most used to the least used.  A Uni-Gram is a single word.
}
\examples{
\dontrun{ 
library(SAoTD)
data <- raw_tweets
TD_Unigram <- Unigram(DataFrame = data)
TD_Unigram
 }                        
}
