% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ViolinPlot}
\alias{ViolinPlot}
\title{Twitter Data Violin Plot}
\usage{
ViolinPlot(DataFrameTidyScores, HT_Topic)
}
\arguments{
\item{DataFrameTidyScores}{DataFrame of Twitter Data that has been tidy'd and scored.}

\item{HT_Topic}{If using hashtag data select:  "hashtag".  If using topic data select:  "topic".}
}
\value{
A ggplot violin plot.
}
\description{
Displays the distribution scores of either hashtag or topic Twitter data.
}
\examples{
\dontrun{ 
library(SAoTD)
data <- raw_tweets
tidy_data <- Tidy(DataFrame = data)
score_data <- Scores(DataFrameTidy = tidy_data, 
                     HT_Topic = "hashtag") 
ht_violin <- ViolinPlot(DataFrameTidyScores = score_data,
                        HT_Topic = "hashtag")
ht_violin

data <- raw_tweets
TD_Topics <- SAoTD::Tweet.Topics(DataFrame = data, 
                                   clusters = 5, 
                                     method = "Gibbs", 
                                   set_seed = 1234, 
                                  num_terms = 10)

TD_Topics <- TD_Topics \%>\% 
  dplyr::mutate(Topic = stringr::str_replace_all(Topic, "^1$", "travel")) \%>\% 
  dplyr::mutate(Topic = stringr::str_replace_all(Topic, "^2$", "recreation")) \%>\% 
  dplyr::mutate(Topic = stringr::str_replace_all(Topic, "^3$", "hiring")) \%>\% 
  dplyr::mutate(Topic = stringr::str_replace_all(Topic, "^4$", "cats")) \%>\% 
  dplyr::mutate(Topic = stringr::str_replace_all(Topic, "^5$", "random"))             
                                         
tidy_data <- Tidy(DataFrame = TD_Topics)
score_data <- Scores(DataFrameTidy = tidy_data, 
                     HT_Topic = "topic") 
topic_violin <- ViolinPlot(DataFrameTidyScores = score_data,
                           HT_Topic = "topic") 
topic_violin                    
}
}
